Added:  find maximun position from 3 floating value

+++ maximumUsingGeneric/src/main/java/com/maximumUsingGeneric/FindMaximumNumber.java
@@ -1,20 +1,20 @@
 package com.maximumUsingGeneric;
-
+//uc2 floating find max
 public class FindMaximumNumber {
 
-    public Integer FindMaxValue(Integer firstNumber, Integer secondNumber, Integer thirdNumber) {
-        Integer max = firstNumber;
-        if (secondNumber.compareTo(max) > 0) {
-            max = secondNumber;
-        }
-        if (thirdNumber.compareTo(max) > 0) {
-            max = thirdNumber;
-        }
-        printMax(max);
-        return max;
-    }
+	public Float FindMaxValue(Float firstNumber, Float secondNumber, Float thirdNumber) {
+		Float max = firstNumber;
+		if (secondNumber.compareTo(max) > 0) {
+			max = secondNumber;
+		}
+		if (thirdNumber.compareTo(max) > 0) {
+			max = thirdNumber;
+		}
+		printMax(max);
+		return max;
+	}
 
-    private void printMax(Integer max) {
-        System.out.println("Largest = " + max);
-    }
+	private void printMax(Float max) {
+		System.out.println("Largest = " + max);
+	}
 }
